chore: split authentication strategies
